+++ CHANGELOG.md
@@ -77,9 +77,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - sdk(programs): Added `PositionModel::update` function.
 - sdk(sdk): Added `decode_anchor_event_with_options` utility function.
 - sdk(js): Added support for `TradeEvent`.
+- sdk(sdk): Added support for `mint_gt_reward` instruction.
 - cli: Added `lp init-lp` command.
 - cli: Added `exchange close-empty-positions` command.
 - cli: Added options to keep or close position.
+- cli: Added `gt mint-reward` command.
 
 ### Changed